Add ApiClient class with login and make_api_request methods
