+++ fetch.py
@@ -11,7 +11,6 @@ Auth has two versions:
 
 import argparse
 from datetime import datetime
-import json
 import os
 import shutil
 import time
@@ -20,87 +19,19 @@ import urllib.request
 import piexif
 from PIL import Image
 
-class FamlyClient:
-    _access_token = None
+from api_client import ApiClient
+
+class FamlyDownloader:
 
     def __init__(self, email, password):
-        self._base = "https://app.famly.co"
         self._pictures_folder = "pictures"
-
-        login_data = self.login(email, password)
-        self._access_token = login_data["data"]["me"]["authenticateWithPassword"]["accessToken"]
-
-    def _request_json(self, method, url, body=None):
-        """Get some JSON and hope we get some back"""
-        b = None
-        if body:
-            b = json.dumps(body).encode("utf-8")
-
-        headers = {"Content-Type": "application/json"}
-        if self._access_token:
-            headers["x-famly-accesstoken"] = self._access_token
-
-        req = urllib.request.Request(
-            url=url, headers=headers, method=method, data=b)
-        try:
-            with urllib.request.urlopen(req) as f:
-                body = f.read().decode("utf-8")
-                if f.status != 200:
-                    raise "B0rked! %" % body
-
-                try:
-                    return json.loads(body)
-                except Exception as e:
-                    return body
-        except urllib.error.HTTPError as e:
-            # The server couldn't fulfill the request
-            print('Error code: ', e.code)
-            print('Response body: ', e.read())
-
-    def _auth_request(self, method, path, body=None, request_params=None):
-        # This should perhaps be split in two functions, one for "classic"
-        # requests and auth, and then one for the "v2"-stuff...
-        if not request_params:
-            request_params = {}
-
-        if "v2" not in path:
-            request_params["accessToken"] = self._access_token
-
-        # Serialize request-params
-        key_values = ["%s=%s" % (k, v) for k, v in request_params.items()]
-
-        return self._request_json(
-            method, self._base + path + "?" + ("&".join(key_values)), body
-        )
-
-    def login(self, email, password):
-        """Get an access token"""
-
-        postBody = {
-            "operationName": "Authenticate",
-            "variables": {
-                "email": email,
-                "password": password,
-                "deviceId": "d2900c00-042d-4db2-a329-798fcd2f152e",
-                "legacy": False,
-            },
-            "query": "mutation Authenticate($email: EmailAddress!, $password: Password!, $deviceId: DeviceId, $legacy: Boolean) {\n  me {\n    authenticateWithPassword(\n      email: $email\n      password: $password\n      deviceId: $deviceId\n      legacy: $legacy\n    ) {\n      ...AuthenticationResult\n      __typename\n    }\n    __typename\n  }\n}\n\nfragment AuthenticationResult on AuthenticationResult {\n  status\n  __typename\n  ... on AuthenticationFailed {\n    status\n    errorDetails\n    errorTitle\n    __typename\n  }\n  ... on AuthenticationSucceeded {\n    accessToken\n    deviceId\n    __typename\n  }\n  ... on AuthenticationChallenged {\n    ...AuthChallenge\n    __typename\n  }\n}\n\nfragment AuthChallenge on AuthenticationChallenged {\n  loginId\n  deviceId\n  expiresAt\n  choices {\n    context {\n      ...UserContextFragment\n      __typename\n    }\n    hmac\n    requiresTwoFactor\n    __typename\n  }\n  person {\n    name {\n      fullName\n      __typename\n    }\n    profileImage {\n      url\n      __typename\n    }\n    __typename\n  }\n  __typename\n}\n\nfragment UserContextFragment on UserContext {\n  id\n  target {\n    __typename\n    ... on PersonContextTarget {\n      person {\n        name {\n          fullName\n          __typename\n        }\n        __typename\n      }\n      children {\n        name {\n          firstName\n          fullName\n          __typename\n        }\n        profileImage {\n          url\n          __typename\n        }\n        __typename\n      }\n      __typename\n    }\n    ... on InstitutionSet {\n      title\n      profileImage {\n        url\n        __typename\n      }\n      __typename\n    }\n  }\n  __typename\n}"
-        }
-
-        return self._request_json(
-            "POST",
-            self._base + "/graphql?Authenticate",
-            body=postBody,
-        )
-
-    def me_me_me(self):
-        """Know about thy self..."""
-        return self._auth_request("GET", "/api/me/me/me")
-
+        self._apiClient = ApiClient()
+        self._apiClient.login(email, password)
+    
     def download_images_by_child_id(self, child_id, first_name):
         """Download images by childId"""
-        imgs = client._auth_request(
-            "GET", "/api/v2/images/tagged", request_params={"childId": child_id}
+        imgs = self._apiClient.make_api_request(
+            "GET", "/api/v2/images/tagged", params={"childId": child_id}
         )
 
         print("Fetching %s images for %s" % (len(imgs), first_name))
@@ -148,17 +79,21 @@ class FamlyClient:
 
 
 if __name__ == "__main__":
+    # Parse arguments
     parser = argparse.ArgumentParser(description="Fetch kids' images from famly.co")
     parser.add_argument("email", help="Auth email")
     parser.add_argument("password", help="Auth password")
     args = parser.parse_args()
-    client = FamlyClient(args.email, args.password)
 
-    my_info = client.me_me_me()
+    # Create the downloader
+    famly_downloader = FamlyDownloader(args.email, args.password)
+
+    my_info = famly_downloader._apiClient.me_me_me()
+
 
     # Current children
     for role in my_info["roles2"]:
-        client.download_images_by_child_id(role["targetId"], role["title"])
+        famly_downloader.download_images_by_child_id(role["targetId"], role["title"])
 
     # Previous children (that's what they call it)
     prev_children = []
@@ -167,5 +102,5 @@ if __name__ == "__main__":
             prev_children = ele["payload"]["children"]
 
     for child in prev_children:
-        client.download_images_by_child_id(
+        famly_downloader.download_images_by_child_id(
             child["childId"], child["name"]["firstName"])
